# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ForgotPasswordResponse": {
				"description": "Forgot password response",
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"success",
							"error"
						]
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ForgotPasswordRequest": {
				"description": "Forgot password request",
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResetPasswordResponse": {
				"description": "Reset password response",
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"success",
							"error"
						]
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"token": {
								"type": "string"
							}
						},
						"required": [
							"token"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResetPasswordRequest": {
				"description": "Reset password request",
				"properties": {
					"token": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"required": [
					"token",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RefreshTokenResponse": {
				"description": "Refresh token response",
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"success",
							"error"
						]
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"refreshToken": {
								"type": "string"
							},
							"token": {
								"type": "string"
							},
							"user": {
								"properties": {
									"avatar": {
										"type": "string"
									},
									"isActive": {
										"type": "boolean"
									},
									"isEmailVerified": {
										"type": "boolean"
									},
									"storeName": {
										"type": "string"
									},
									"storeId": {
										"type": "string"
									},
									"role": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"name": {
										"type": "string"
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"isActive",
									"isEmailVerified",
									"role",
									"email",
									"id"
								],
								"type": "object"
							}
						},
						"required": [
							"refreshToken",
							"token",
							"user"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RefreshTokenRequest": {
				"description": "Refresh token request",
				"properties": {
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetProfileResponse": {
				"description": "Get profile response",
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"success",
							"error"
						]
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"user": {
								"properties": {
									"lastLoginAt": {
										"type": "string",
										"format": "date-time"
									},
									"createdAt": {
										"type": "string",
										"format": "date-time"
									},
									"totalSpent": {
										"type": "number",
										"format": "double"
									},
									"totalOrdersCount": {
										"type": "number",
										"format": "double"
									},
									"preferences": {},
									"addresses": {
										"items": {},
										"type": "array"
									},
									"avatar": {
										"type": "string"
									},
									"isActive": {
										"type": "boolean"
									},
									"isEmailVerified": {
										"type": "boolean"
									},
									"storeName": {
										"type": "string"
									},
									"storeId": {
										"type": "string"
									},
									"role": {
										"type": "string"
									},
									"defaultAddress": {
										"nullable": true
									},
									"dateOfBirth": {
										"type": "string"
									},
									"gender": {
										"type": "string"
									},
									"country": {
										"type": "string"
									},
									"phoneNumber": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"fullName": {
										"type": "string"
									},
									"id": {
										"type": "string"
									}
								},
								"required": [
									"createdAt",
									"isActive",
									"isEmailVerified",
									"role",
									"defaultAddress",
									"dateOfBirth",
									"gender",
									"country",
									"phoneNumber",
									"email",
									"fullName",
									"id"
								],
								"type": "object"
							}
						},
						"required": [
							"user"
						],
						"type": "object"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProfileResponse": {
				"description": "Update profile response",
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"success",
							"error"
						]
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"updatedFields": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"user": {}
						},
						"required": [
							"updatedFields",
							"user"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ChangePasswordRequest": {
				"description": "Change password request",
				"properties": {
					"currentPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"required": [
					"currentPassword",
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Cartaisy API",
		"version": "1.0.0",
		"description": "E-commerce backend API with Shopify integration"
	},
	"paths": {
...
		"/auth/forgot-password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "Password reset email sent if account exists",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForgotPasswordResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Request password reset",
				"summary": "Send password reset email",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Email address for password reset",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ForgotPasswordRequest",
								"description": "Email address for password reset"
							}
						}
					}
				}
			}
		},
		"/auth/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Password reset successful",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResetPasswordResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request - Invalid or expired token"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Reset password with token",
				"summary": "Reset password using reset token from email",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Reset token and new password",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResetPasswordRequest",
								"description": "Reset token and new password"
							}
						}
					}
				}
			}
		},
		"/auth/refresh-token": {
			"post": {
				"operationId": "RefreshToken",
				"responses": {
					"200": {
						"description": "Token refreshed successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RefreshTokenResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request - Refresh token required"
					},
					"401": {
						"description": "Unauthorized - Invalid or expired refresh token"
					},
					"403": {
						"description": "Forbidden - Account deactivated"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Refresh access token",
				"summary": "Get new access token using refresh token",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Refresh token",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RefreshTokenRequest",
								"description": "Refresh token"
							}
						}
					}
				}
			}
		},
		"/auth/profile": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Profile retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetProfileResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Get current user profile",
				"summary": "Get authenticated user's profile data",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"patch": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Profile updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UpdateProfileResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Unauthorized"
					},
					"404": {
						"description": "User not found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Update user profile",
				"summary": "Update authenticated user's profile",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Profile fields to update",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProfileRequest",
								"description": "Profile fields to update"
							}
						}
					}
				}
			}
		},
		"/auth/change-password": {
			"post": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Password changed successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ChangePasswordResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"404": {
						"description": "User not found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Change password",
				"summary": "Change authenticated user's password",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Current password and new password",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordRequest",
								"description": "Current password and new password"
							}
						}
					}
				}
			}
		},
		"/auth/account": {
			"delete": {
				"operationId": "DeleteAccount",
				"responses": {
					"200": {
						"description": "Account deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteAccountResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Unauthorized"
					},
					"404": {
						"description": "User not found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Delete user account",
				"summary": "Soft delete user account (requires password verification)",
				"tags": [
					"Authentication"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Password for verification",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteAccountRequest",
								"description": "Password for verification"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/api/v1",
			"description": "Development server"
		}
	]
}